+++ apps/web/src/app/HR/_components/dashboard/AuditLogRuleDecisionDialog.tsx
@@ -1,9 +1,17 @@
 "use client";
 
-import { useEffect } from "react";
-import { Asterisk, ChevronRight, X } from "lucide-react";
+import { Asterisk } from "lucide-react";
 
 import { formatDetailDateTimeWithAt } from "@/app/(hr)/requests/components/request-detail-formatters";
+import {
+  Card,
+  ChangeSummaryCard,
+  DecisionTextCard,
+  Field,
+  Section,
+} from "./audit-log-dialog-primitives";
+import { AuditLogDialogShell } from "./audit-log-dialog-shell";
+import { getRuleDecisionDialogState } from "./audit-log-dialog-view-models";
 import type { AuditLogEntry } from "./audit-log-types";
 
 type AuditLogRuleDecisionDialogProps = {
@@ -17,294 +25,144 @@ export default function AuditLogRuleDecisionDialog({
 }: AuditLogRuleDecisionDialogProps) {
   const detail = entry.ruleApprovalDetail;
 
-  useEffect(() => {
-    if (!detail) return;
-
-    const previousBodyOverflow = document.body.style.overflow;
-    const previousBodyOverscrollBehavior = document.body.style.overscrollBehavior;
-
-    document.body.style.overflow = "hidden";
-    document.body.style.overscrollBehavior = "none";
-
-    return () => {
-      document.body.style.overflow = previousBodyOverflow;
-      document.body.style.overscrollBehavior = previousBodyOverscrollBehavior;
-    };
-  }, [detail]);
-
   if (!detail) {
     return null;
   }
 
-  const isApproved = entry.result === "Approved";
-  const isArchived = isApproved && detail.actionType === "delete";
-  const dialogEventLabel = isArchived ? "Rule Archived" : entry.event;
-  const resultBadgeLabel = isArchived ? "Archived" : entry.result;
-  const changeSummaryItems =
-    detail.actionType === "update"
-      ? [
-          detail.previousRequirementValue &&
-          detail.previousRequirementValue !== detail.requirementValue
-            ? {
-                currentValue: detail.requirementValue,
-                label: "Requirement Value",
-                previousValue: detail.previousRequirementValue,
-              }
-            : null,
-          detail.previousBlockingMessage &&
-          detail.previousBlockingMessage !== detail.blockingMessage
-            ? {
-                currentValue: detail.blockingMessage,
-                label: "Blocking Message",
-                previousValue: detail.previousBlockingMessage,
-              }
-            : null,
-        ].filter(
-          (
-            item,
-          ): item is {
-            currentValue: string;
-            label: string;
-            previousValue: string;
-          } => Boolean(item),
-        )
-      : [];
-  const showChangeSummary = changeSummaryItems.length > 0;
-  const showNotesSection = !isArchived && (!isApproved || detail.actionType !== "update");
-  const showSubmissionDetails = !isArchived;
-  const showTargetBenefits = !isArchived;
-  const notesFallback = isApproved
-    ? "No decision notes were saved for this request."
-    : "No rejection reason was saved for this request.";
+  const state = getRuleDecisionDialogState(entry, detail);
 
   return (
-    <div
-      className="fixed inset-0 z-[70] flex items-start justify-center overflow-hidden bg-black/50 px-4 py-6"
-      onClick={(event) => {
-        if (event.target === event.currentTarget) onClose();
-      }}
+    <AuditLogDialogShell
+      compact={state.isArchived}
+      icon={Asterisk}
+      onClose={onClose}
+      title={state.dialogEventLabel}
     >
-      <div
-        aria-modal="true"
-        className={`hidden-scrollbar relative mx-auto flex max-h-[calc(100vh-48px)] w-full flex-col items-start gap-4 overflow-y-auto overscroll-none rounded-[10px] border border-[#E5E5E5] bg-white shadow-[0px_10px_15px_-3px_rgba(0,0,0,0.1),0px_4px_6px_-4px_rgba(0,0,0,0.1)] ${
-          isArchived ? "max-w-[512px] px-[23.8px] py-[23.8px]" : "max-w-[506px] px-[25px] py-[30px]"
-        }`}
-        role="dialog"
-      >
-        <button
-          aria-label="Close dialog"
-          className="absolute top-5 right-5 rounded-[8px] p-2 text-[#737373] transition-colors hover:bg-[#F5F5F5]"
-          onClick={onClose}
-          type="button"
-        >
-          <X className="h-5 w-5" />
-        </button>
-
-        <div className="flex w-full flex-col gap-2 pr-10">
-          <div className="flex items-center gap-2">
-            <Asterisk className="h-4 w-4 text-[#737373]" />
-            <h2 className="text-[18px] leading-[18px] font-semibold text-[#0A0A0A]">
-              {dialogEventLabel}
-            </h2>
+      <Section title="Action Overview">
+        <Card>
+          <div className="grid gap-4 sm:grid-cols-2">
+            <Field label="Timestamp" value={formatDetailDateTimeWithAt(entry.occurredAt)} />
+            <Field label="Event Type" value={state.dialogEventLabel} />
+            <ResultBadge isApproved={state.isApproved} isArchived={state.isArchived} label={state.resultBadgeLabel} />
+            <PersonField label="Performed By" person={entry.performedBy} />
+            {!state.isArchived ? <PersonField label="Reviewed By" person={detail.reviewedBy} /> : null}
           </div>
-          <p className="text-[14px] leading-5 text-[#737373]">Full details for this action.</p>
-        </div>
+        </Card>
+      </Section>
 
-        <div className="flex w-full flex-col gap-6">
-          <Section title="Action Overview">
-            <Card>
-              <div className="grid gap-4 sm:grid-cols-2">
-                <Field label="Timestamp" value={formatDetailDateTimeWithAt(entry.occurredAt)} />
-                <Field label="Event Type" value={dialogEventLabel} />
-                <div className="flex flex-col gap-[6px]">
-                  <span className="text-[12px] leading-4 text-[#737373]">Result</span>
-                  <span
-                    className={`inline-flex w-fit items-center justify-center rounded-[8px] px-[7.8px] py-[1.8px] text-[12px] leading-4 font-medium ${
-                      isArchived
-                        ? "bg-[#F3F4F6] text-[#1E2939]"
-                        : isApproved
-                        ? "bg-[#DCFCE7] text-[#016630]"
-                        : "bg-[#FFE2E2] text-[#9F0712]"
-                    }`}
-                  >
-                    {resultBadgeLabel}
-                  </span>
-                </div>
-                <div className="flex flex-col gap-1">
-                  <span className="text-[12px] leading-4 text-[#737373]">Performed By</span>
-                  <span className="text-[14px] leading-5 font-medium text-[#0A0A0A]">
-                    {entry.performedBy.name}
-                  </span>
-                  <span className="text-[12px] leading-4 text-[#737373]">{entry.performedBy.role}</span>
-                </div>
-                {!isArchived ? (
-                  <div className="flex flex-col gap-1">
-                    <span className="text-[12px] leading-4 text-[#737373]">Reviewed By</span>
-                    <span className="text-[14px] leading-5 font-medium text-[#0A0A0A]">
-                      {detail.reviewedBy.name}
-                    </span>
-                    <span className="text-[12px] leading-4 text-[#737373]">{detail.reviewedBy.role}</span>
-                  </div>
-                ) : null}
-              </div>
-            </Card>
-          </Section>
+      <Section title="Rule Details">
+        <Card>
+          <div className="flex flex-col gap-4">
+            <div className="grid gap-4 sm:grid-cols-2">
+              <Field label="Rule Name" value={detail.name} />
+              <Field label="Category" value={detail.category} />
+              <Field label="Source Field" value={detail.sourceField} />
+              <Field label="Requirement Value" value={detail.requirementValue} />
+            </div>
+            <Field label="Blocking Message" value={detail.blockingMessage} />
+            {state.showTargetBenefits ? <TargetBenefits benefits={detail.targetBenefits} /> : null}
+          </div>
+        </Card>
+      </Section>
 
-          <Section title="Rule Details">
-            <Card>
-              <div className="flex flex-col gap-4">
-                <div className="grid gap-4 sm:grid-cols-2">
-                  <Field label="Rule Name" value={detail.name} />
-                  <Field label="Category" value={detail.category} />
-                  <Field label="Source Field" value={detail.sourceField} />
-                  <Field label="Requirement Value" value={detail.requirementValue} />
-                </div>
-                <Field label="Blocking Message" value={detail.blockingMessage} />
-                {showTargetBenefits ? (
-                  <div className="flex flex-col gap-2">
-                    <span className="text-[12px] leading-4 text-[#737373]">Target Benefits</span>
-                    <div className="flex flex-wrap gap-1.5">
-                      {detail.targetBenefits.length > 0 ? (
-                        detail.targetBenefits.map((benefit) => (
-                          <span
-                            className="inline-flex items-center justify-center rounded-[8px] border border-[#E5E5E5] px-[7.8px] py-[2px] text-[12px] leading-4 font-medium text-[#0A0A0A]"
-                            key={benefit}
-                          >
-                            {benefit}
-                          </span>
-                        ))
-                      ) : (
-                        <span className="text-[14px] leading-5 text-[#737373]">No linked benefits</span>
-                      )}
-                    </div>
-                  </div>
-                ) : null}
-              </div>
-            </Card>
-          </Section>
+      {state.showSubmissionDetails ? (
+        <Section title="Submission Details">
+          <Card>
+            <div className="grid gap-4 sm:grid-cols-2">
+              <Field label="Submitted By" value={detail.submittedBy.name} />
+              <Field label="Submitted At" value={formatDetailDateTimeWithAt(detail.submittedAt)} />
+              <Field
+                label={state.isApproved ? "Approved At" : "Rejected At"}
+                value={formatDetailDateTimeWithAt(detail.decisionAt)}
+              />
+            </div>
+          </Card>
+        </Section>
+      ) : null}
 
-          {showSubmissionDetails ? (
-            <Section title="Submission Details">
-              <Card>
-                <div className="grid gap-4 sm:grid-cols-2">
-                  <Field label="Submitted By" value={detail.submittedBy.name} />
-                  <Field label="Submitted At" value={formatDetailDateTimeWithAt(detail.submittedAt)} />
-                  <Field
-                    label={isApproved ? "Approved At" : "Rejected At"}
-                    value={formatDetailDateTimeWithAt(detail.decisionAt)}
-                  />
-                </div>
-              </Card>
-            </Section>
-          ) : null}
+      {state.showChangeSummary ? (
+        <Section title="Change Summary">
+          <div className="flex flex-col gap-2">
+            {state.changeSummaryItems.map((item) => (
+              <ChangeSummaryCard {...item} key={item.label} />
+            ))}
+          </div>
+        </Section>
+      ) : null}
 
-          {showChangeSummary ? (
-            <Section title="Change Summary">
-              <div className="flex flex-col gap-2">
-                {changeSummaryItems.map((item) => (
-                  <ChangeSummaryCard
-                    currentValue={item.currentValue}
-                    key={item.label}
-                    label={item.label}
-                    previousValue={item.previousValue}
-                  />
-                ))}
-              </div>
-            </Section>
-          ) : null}
-
-          {showNotesSection ? (
-            <Section title={isApproved ? "Decision Notes" : "Rejection Reason"}>
-              <div
-                className={`w-full rounded-[10px] border px-[15.8px] py-[15.8px] ${
-                  isApproved
-                    ? "border-[#E5E5E5] bg-[rgba(245,245,245,0.3)]"
-                    : "border-[#FFC9C9] bg-[#FEF2F2]"
-                }`}
-              >
-                <p
-                  className={`text-[14px] leading-5 ${
-                    isApproved ? "text-[#0A0A0A]" : "text-[#C10007]"
-                  }`}
-                >
-                  {detail.reviewComment?.trim() || notesFallback}
-                </p>
-              </div>
-            </Section>
-          ) : null}
-        </div>
-      </div>
-    </div>
+      {state.showNotesSection ? (
+        <Section title={state.isApproved ? "Decision Notes" : "Rejection Reason"}>
+          <DecisionTextCard
+            isPositive={state.isApproved}
+            text={detail.reviewComment?.trim() || state.notesFallback}
+          />
+        </Section>
+      ) : null}
+    </AuditLogDialogShell>
   );
 }
 
-function Section({
-  children,
-  title,
-}: {
-  children: React.ReactNode;
-  title: string;
-}) {
-  return (
-    <section className="flex w-full flex-col gap-3">
-      <h3 className="text-[14px] leading-5 font-semibold text-[#0A0A0A]">{title}</h3>
-      {children}
-    </section>
-  );
-}
-
-function Card({
-  children,
-  className = "",
-}: {
-  children: React.ReactNode;
-  className?: string;
-}) {
-  return (
-    <div className={`w-full rounded-[10px] border border-[#E5E5E5] bg-[rgba(245,245,245,0.3)] px-[15.8px] py-[15.8px] ${className}`.trim()}>
-      {children}
-    </div>
-  );
-}
-
-function ChangeSummaryCard({
-  currentValue,
+function PersonField({
   label,
-  previousValue,
-}: {
-  currentValue: string;
-  label: string;
-  previousValue: string;
-}) {
-  return (
-    <Card className="px-[11.8px] py-[11.8px]">
-      <div className="flex flex-col gap-2">
-        <span className="text-[12px] leading-4 text-[#737373]">{label}</span>
-        <div className="flex items-center gap-3">
-          <div className="flex-1 rounded-[4px] border border-[#FFC9C9] bg-[#FEF2F2] px-[7.8px] py-[7.8px] text-[14px] leading-5 text-[#C10007] line-through">
-            {previousValue}
-          </div>
-          <ChevronRight className="h-4 w-4 shrink-0 text-[#737373]" />
-          <div className="flex-1 rounded-[4px] border border-[#B9F8CF] bg-[#F0FDF4] px-[7.8px] py-[7.8px] text-[14px] leading-5 text-[#008236]">
-            {currentValue}
-          </div>
-        </div>
-      </div>
-    </Card>
-  );
-}
-
-function Field({
-  label,
-  value,
+  person,
 }: {
   label: string;
-  value: React.ReactNode;
+  person: { name: string; role: string };
 }) {
   return (
     <div className="flex flex-col gap-1">
       <span className="text-[12px] leading-4 text-[#737373]">{label}</span>
-      <span className="text-[14px] leading-5 font-medium text-[#0A0A0A]">{value}</span>
+      <span className="text-[14px] leading-5 font-medium text-[#0A0A0A]">{person.name}</span>
+      <span className="text-[12px] leading-4 text-[#737373]">{person.role}</span>
+    </div>
+  );
+}
+
+function ResultBadge({
+  isApproved,
+  isArchived,
+  label,
+}: {
+  isApproved: boolean;
+  isArchived: boolean;
+  label: string;
+}) {
+  return (
+    <div className="flex flex-col gap-[6px]">
+      <span className="text-[12px] leading-4 text-[#737373]">Result</span>
+      <span
+        className={`inline-flex w-fit items-center justify-center rounded-[8px] px-[7.8px] py-[1.8px] text-[12px] leading-4 font-medium ${
+          isArchived
+            ? "bg-[#F3F4F6] text-[#1E2939]"
+            : isApproved
+              ? "bg-[#DCFCE7] text-[#016630]"
+              : "bg-[#FFE2E2] text-[#9F0712]"
+        }`}
+      >
+        {label}
+      </span>
+    </div>
+  );
+}
+
+function TargetBenefits({ benefits }: { benefits: string[] }) {
+  return (
+    <div className="flex flex-col gap-2">
+      <span className="text-[12px] leading-4 text-[#737373]">Target Benefits</span>
+      <div className="flex flex-wrap gap-1.5">
+        {benefits.length > 0 ? (
+          benefits.map((benefit) => (
+            <span
+              className="inline-flex items-center justify-center rounded-[8px] border border-[#E5E5E5] px-[7.8px] py-[2px] text-[12px] leading-4 font-medium text-[#0A0A0A]"
+              key={benefit}
+            >
+              {benefit}
+            </span>
+          ))
+        ) : (
+          <span className="text-[14px] leading-5 text-[#737373]">No linked benefits</span>
+        )}
+      </div>
     </div>
   );
 }
