+++ apps/web/src/app/HR/_components/dashboard/audit-log-builder-details.ts
@@ -1,0 +1,5 @@
+export {
+  buildApprovalBenefitRequestDetail,
+  buildBenefitApprovalDetail,
+} from "./audit-log-builder-benefit-details";
+export { buildRuleApprovalDetail } from "./audit-log-builder-rule-details";
