+++ apps/web/src/shared/auth/get-employee-record-by-email.ts
@@ -1,11 +1,11 @@
-/* eslint-disable max-lines */
 import "server-only";
 
 import { getCloudflareContext } from "@opennextjs/cloudflare";
-
-const DEFAULT_GRAPHQL_ENDPOINT =
-  "https://ebms-backend.b94889340.workers.dev/graphql";
-const DEFAULT_LOCAL_GRAPHQL_ENDPOINT = "http://127.0.0.1:8787/graphql";
+import {
+  fetchEmployeeRecordFromEmployeesQuery,
+  fetchEmployeeRecordFromGraphql,
+  getGraphqlEndpoints,
+} from "./employee-record-graphql";
 
 export type EmployeeRecord = {
   department: string;
@@ -20,24 +20,6 @@ export type EmployeeRecord = {
   responsibilityLevel: number;
 };
 
-type EmployeeByEmailResponse = {
-  data?: {
-    employeeByEmail: EmployeeRecord | null;
-  };
-  errors?: Array<{
-    message: string;
-  }>;
-};
-
-type EmployeesResponse = {
-  data?: {
-    employees: EmployeeRecord[];
-  };
-  errors?: Array<{
-    message: string;
-  }>;
-};
-
 type D1Like = {
   prepare: (query: string) => {
     bind: (...values: unknown[]) => {
@@ -46,47 +28,64 @@ type D1Like = {
   };
 };
 
-function getGraphqlEndpoint() {
-  const endpoint =
-    process.env.GRAPHQL_ENDPOINT ?? process.env.NEXT_PUBLIC_GRAPHQL_ENDPOINT;
+export async function getEmployeeRecordByEmail(email: string) {
+  const normalizedEmail = email.trim().toLowerCase();
+  const employeeFromD1 = await fetchEmployeeRecordFromD1(normalizedEmail);
+  let resolvedNoMatchCount = 0;
+  let lookupFailureCount = 0;
 
-  if (endpoint) {
-    return endpoint;
+  if (employeeFromD1) {
+    return employeeFromD1;
   }
 
-  if (process.env.NODE_ENV === "development") {
-    return DEFAULT_LOCAL_GRAPHQL_ENDPOINT;
+  for (const endpoint of getGraphqlEndpoints()) {
+    try {
+      const employee = await fetchEmployeeRecordFromGraphql(endpoint, normalizedEmail);
+      if (employee) {
+        return employee;
+      }
+
+      resolvedNoMatchCount += 1;
+      console.warn("[auth] employeeByEmail returned no match.", {
+        email: normalizedEmail,
+        endpoint,
+      });
+    } catch (graphqlError) {
+      try {
+        const employee = await fetchEmployeeRecordFromEmployeesQuery(endpoint, normalizedEmail);
+        if (employee) {
+          return employee;
+        }
+
+        resolvedNoMatchCount += 1;
+        console.warn("[auth] employees query returned no match.", {
+          email: normalizedEmail,
+          endpoint,
+        });
+      } catch (employeesError) {
+        lookupFailureCount += 1;
+        console.error("[auth] Employee lookup failed for endpoint.", {
+          email: normalizedEmail,
+          endpoint,
+          employeesError:
+            employeesError instanceof Error
+              ? employeesError.message
+              : String(employeesError),
+          graphqlError:
+            graphqlError instanceof Error ? graphqlError.message : String(graphqlError),
+        });
+      }
+    }
   }
 
-  return DEFAULT_GRAPHQL_ENDPOINT;
-}
-
-function getGraphqlEndpoints() {
-  return [...new Set([getGraphqlEndpoint(), DEFAULT_GRAPHQL_ENDPOINT])];
-}
-
-async function postGraphql<T>(
-  endpoint: string,
-  query: string,
-  variables: Record<string, string>,
-) {
-  const response = await fetch(endpoint, {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    cache: "no-store",
-    body: JSON.stringify({
-      query,
-      variables,
-    }),
-  });
-
-  if (!response.ok) {
-    throw new Error(`Failed to load employee access (${response.status}).`);
+  if (lookupFailureCount > 0) {
+    throw new Error("Failed to load employee access.");
+  }
+  if (resolvedNoMatchCount > 0) {
+    return null;
   }
 
-  return (await response.json()) as T;
+  throw new Error("Failed to resolve employee access.");
 }
 
 async function fetchEmployeeRecordFromD1(email: string) {
@@ -126,141 +125,4 @@ async function fetchEmployeeRecordFromD1(email: string) {
     });
     return null;
   }
-}
-
-async function fetchEmployeeRecordFromGraphql(endpoint: string, email: string) {
-  const payload = await postGraphql<EmployeeByEmailResponse>(
-    endpoint,
-    `
-      query EmployeeByEmail($email: String!) {
-        employeeByEmail(email: $email) {
-          id
-          name
-          email
-          department
-          employmentStatus
-          hireDate
-          position
-          responsibilityLevel
-          okrSubmitted
-          lateArrivalCount30Days: lateArrivalCount
-        }
-      }
-    `,
-    { email },
-  );
-
-  if (payload.errors?.length) {
-    throw new Error(payload.errors[0]?.message ?? "Failed to load employee.");
-  }
-
-  return payload.data?.employeeByEmail ?? null;
-}
-
-async function fetchEmployeeRecordFromEmployeesQuery(
-  endpoint: string,
-  email: string,
-) {
-  const payload = await postGraphql<EmployeesResponse>(
-    endpoint,
-    `
-      query Employees {
-        employees {
-          id
-          name
-          email
-          department
-          employmentStatus
-          hireDate
-          position
-          responsibilityLevel
-          okrSubmitted
-          lateArrivalCount30Days: lateArrivalCount
-        }
-      }
-    `,
-    { email },
-  );
-
-  if (payload.errors?.length) {
-    throw new Error(payload.errors[0]?.message ?? "Failed to load employee.");
-  }
-
-  return (
-    payload.data?.employees.find(
-      (employee) => employee.email.trim().toLowerCase() === email,
-    ) ?? null
-  );
-}
-
-export async function getEmployeeRecordByEmail(email: string) {
-  const normalizedEmail = email.trim().toLowerCase();
-  const employeeFromD1 = await fetchEmployeeRecordFromD1(normalizedEmail);
-  let resolvedNoMatchCount = 0;
-  let lookupFailureCount = 0;
-
-  if (employeeFromD1) {
-    return employeeFromD1;
-  }
-
-  for (const endpoint of getGraphqlEndpoints()) {
-    try {
-      const employee = await fetchEmployeeRecordFromGraphql(
-        endpoint,
-        normalizedEmail,
-      );
-
-      if (employee) {
-        return employee;
-      }
-
-      resolvedNoMatchCount += 1;
-      console.warn("[auth] employeeByEmail returned no match.", {
-        email: normalizedEmail,
-        endpoint,
-      });
-    } catch (graphqlError) {
-      try {
-        const employee = await fetchEmployeeRecordFromEmployeesQuery(
-          endpoint,
-          normalizedEmail,
-        );
-
-        if (employee) {
-          return employee;
-        }
-
-        resolvedNoMatchCount += 1;
-        console.warn("[auth] employees query returned no match.", {
-          email: normalizedEmail,
-          endpoint,
-        });
-      } catch (employeesError) {
-        lookupFailureCount += 1;
-        console.error("[auth] Employee lookup failed for endpoint.", {
-          email: normalizedEmail,
-          endpoint,
-          employeesError:
-            employeesError instanceof Error
-              ? employeesError.message
-              : String(employeesError),
-          graphqlError:
-            graphqlError instanceof Error
-              ? graphqlError.message
-              : String(graphqlError),
-        });
-        continue;
-      }
-    }
-  }
-
-  if (lookupFailureCount > 0) {
-    throw new Error("Failed to load employee access.");
-  }
-
-  if (resolvedNoMatchCount > 0) {
-    return null;
-  }
-
-  throw new Error("Failed to resolve employee access.");
 }
